Added a confirm prompt for safety

+++ MPI Helper Scripts/copyToWorkers.sh
@@ -11,9 +11,19 @@ echo ' '
 read -p 'File(s): ' files
 echo ' '
 
-# Copy Files
-for node in node1 node2 node3 node4 node5; do
-    scp $files $node:/home/mpi/;
-done
+# Confirm
 
+echo 'File(s) to copy: ' files
+echo ' '
+read -p 'Is this correct? [Y/N]: ' -n 1 -r
+
+# If info is correct, run the script
+if  [[ $REPLY =~ ^[Yy]$ ]]
+then 
+        # Copy Files
+        for node in node1 node2 node3 node4 node5; do
+            scp $files $node:/home/mpi/;
+        done
+fi
+echo ' '
 echo 'END SCRIPT'
